Change fn_8018F3BC return type from u8 to int (removes unnecessary clrlwi zero-extends in callers)
fn_80191FD4: 83.48% -> 86.50% (7 extra clrlwi instructions eliminated)
fn_8018F3BC: remains 100%

+++ src/melee/gm/gm_18A5.c
@@ -1359,7 +1359,7 @@ int fn_8018F310(int arg0)
 
 #pragma push
 #pragma auto_inline off
-u8 fn_8018F3BC(s32 arg0)
+int fn_8018F3BC(s32 arg0)
 {
     return lbl_803D9D20.x59[arg0];
 }
